# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Aerospike Backup Service",
        "title": "Backup Service REST API Specification",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/v1/config/storage": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Reads all storage from the configuration.",
                "operationId": "readAllStorage",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/model.Storage"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/config/storage/{name}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Reads a specific storage from the configuration given its name.",
                "operationId": "readStorage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name of the storage",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Storage"
                        }
                    },
                    "404": {
                        "description": "The specified storage could not be found.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Updates an existing storage in the configuration.",
                "operationId": "updateStorage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "storage name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "backup storage",
                        "name": "storage",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Storage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Configuration"
                ],
                "summary": "Adds a storage to the config.",
                "operationId": "addStorage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "storage name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "backup storage",
                        "name": "storage",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Storage"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Deletes a storage from the configuration by name.",
                "operationId": "deleteStorage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "storage name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/restore/full": {
            "post": {
                "description": "Specify the directory parameter for the full backup restore.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Restore"
                ],
                "summary": "Trigger an asynchronous full restore operation.",
                "operationId": "restoreFull",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RestoreRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Job ID",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/restore/incremental": {
            "post": {
                "description": "Specify the file parameter to restore from an incremental backup file.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Restore"
                ],
                "summary": "Trigger an asynchronous incremental restore operation.",
                "operationId": "restoreIncremental",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RestoreRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Job ID",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/restore/status/{jobId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restore"
                ],
                "summary": "Retrieve status for a restore job.",
                "operationId": "restoreStatus",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Job ID to retrieve the status",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job status",
                        "schema": {
                            "$ref": "#/definitions/model.RestoreJobStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/restore/timestamp": {
            "post": {
                "description": "Restores backup from given point in time",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Restore"
                ],
                "summary": "Trigger an asynchronous restore operation to specific point in time.",
                "operationId": "restoreTimestamp",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RestoreTimestampRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Job ID",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.AerospikeCluster": {
            "description": "AerospikeCluster represents the configuration for an Aerospike cluster for backup.",
            "type": "object",
            "properties": {
                "conn-timeout": {
                    "description": "The connection timeout in milliseconds.",
                    "type": "integer"
                },
                "credentials": {
                    "description": "The authentication details to the Aerospike cluster.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Credentials"
                        }
                    ]
                },
                "label": {
                    "description": "The cluster name.",
                    "type": "string"
                },
                "seed-nodes": {
                    "description": "The seed nodes details.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SeedNode"
                    }
                },
                "tls": {
                    "description": "The cluster TLS configuration.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.TLS"
                        }
                    ]
                },
                "use-services-alternate": {
                    "description": "Whether should use \"services-alternate\" instead of \"services\" in info request during cluster tending.",
                    "type": "boolean"
                }
            }
        },
...
        "model.JobStatus": {
            "type": "string",
            "enum": [
                "Running",
                "Done",
                "Failed"
            ],
            "x-enum-varnames": [
                "JobStatusRunning",
                "JobStatusDone",
                "JobStatusFailed"
            ]
        },
...
        "model.RestoreJobStatus": {
            "description": "RestoreJobStatus represents a restore job status.",
            "type": "object",
            "properties": {
                "error": {},
                "existed-records": {
                    "type": "integer"
                },
                "expired-records": {
                    "type": "integer"
                },
                "fresher-records": {
                    "type": "integer"
                },
                "ignored-records": {
                    "type": "integer"
                },
                "index-count": {
                    "type": "integer"
                },
                "inserted-records": {
                    "type": "integer"
                },
                "skipped-records": {
                    "type": "integer"
                },
                "status": {
                    "enum": [
                        "Running",
                        "Done",
                        "Failed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.JobStatus"
                        }
                    ]
                },
                "total-bytes": {
                    "type": "integer"
                },
                "total-records": {
                    "type": "integer"
                },
                "udf-count": {
                    "type": "integer"
                }
            }
        },
...
        "model.RestorePolicy": {
            "description": "RestorePolicy represents a policy for the restore operation.",
            "type": "object",
            "properties": {
                "bandwidth": {
                    "description": "Throttles read operations from the backup file(s) to not exceed the given I/O bandwidth\nin MiB/s and its database write operations to not exceed the given number of transactions\nper second.",
                    "type": "integer"
                },
                "batch-size": {
                    "description": "The max allowed number of records per an async batch write call.\nDefault is 128 with batch writes enabled, or 16 without batch writes.",
                    "type": "integer"
                },
                "bin-list": {
                    "description": "The bins to restore (optional, an empty list implies restoring all bins).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "disable-batch-writes": {
                    "description": "Disables the use of batch writes when restoring records to the Aerospike cluster.\nBy default, the cluster is checked for batch write support.",
                    "type": "boolean"
                },
                "max-async-batches": {
                    "description": "The max number of outstanding async record batch write calls at a time.",
                    "type": "integer"
                },
                "namespace": {
                    "description": "Namespace details for the restore operation.\nBy default, the data is restored to the namespace from which it was taken.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.RestoreNamespace"
                        }
                    ]
                },
                "no-generation": {
                    "description": "Records from backups take precedence. This option disables the generation check.\nWith this option, records from the backup always overwrite records that already exist in\nthe namespace, regardless of generation numbers.",
                    "type": "boolean"
                },
                "no-indexes": {
                    "description": "Do not restore any secondary index definitions.",
                    "type": "boolean"
                },
                "no-records": {
                    "description": "Do not restore any record data (metadata or bin data).\nBy default, record data, secondary index definitions, and UDF modules\nwill be restored.",
                    "type": "boolean"
                },
                "no-udfs": {
                    "description": "Do not restore any UDF modules.",
                    "type": "boolean"
                },
                "parallel": {
                    "description": "The number of client threads to spawn for writing to the cluster.",
                    "type": "integer"
                },
                "replace": {
                    "description": "Replace records. This controls how records from the backup overwrite existing records in\nthe namespace. By default, restoring a record from a backup only replaces the bins\ncontained in the backup; all other bins of an existing record remain untouched.",
                    "type": "boolean"
                },
                "set-list": {
                    "description": "The sets to restore (optional, an empty list implies restoring all sets).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "timeout": {
                    "description": "Timeout (ms) for Aerospike commands to write records, create indexes and create UDFs.",
                    "type": "integer"
                },
                "tps": {
                    "description": "Throttles read operations from the backup file(s) to not exceed the given I/O bandwidth\nin MiB/s and its database write operations to not exceed the given number of transactions\nper second.",
                    "type": "integer"
                },
                "unique": {
                    "description": "Existing records take precedence. With this option, only records that do not exist in\nthe namespace are restored, regardless of generation numbers. If a record exists in\nthe namespace, the record from the backup is ignored.",
                    "type": "boolean"
                }
            }
        },
        "model.RestoreRequest": {
            "description": "RestoreRequest represents a restore operation request.",
            "type": "object",
            "properties": {
                "destination": {
                    "$ref": "#/definitions/model.AerospikeCluster"
                },
                "policy": {
                    "$ref": "#/definitions/model.RestorePolicy"
                },
                "secret-agent": {
                    "$ref": "#/definitions/model.SecretAgent"
                },
                "source": {
                    "$ref": "#/definitions/model.Storage"
                }
            }
        },
        "model.RestoreTimestampRequest": {
            "description": "RestoreTimestampRequest represents a restore by timestamp operation request.",
            "type": "object",
            "properties": {
                "destination": {
                    "description": "The details of the Aerospike destination cluster.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AerospikeCluster"
                        }
                    ]
                },
                "policy": {
                    "description": "Restore policy to use in the operation.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.RestorePolicy"
                        }
                    ]
                },
                "routine": {
                    "description": "The backup routine name.",
                    "type": "string"
                },
                "secret-agent": {
                    "description": "Secret Agent configuration (optional).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.SecretAgent"
                        }
                    ]
                },
                "time": {
                    "description": "Required epoch time for recovery. The closest backup before the timestamp will be applied.",
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "model.SecretAgent": {
            "description": "SecretAgent represents the configuration of an Aerospike Secret Agent for a backup/restore operation.",
            "type": "object",
            "properties": {
                "address": {
                    "description": "Address of the Secret Agent.",
                    "type": "string"
                },
                "port": {
                    "description": "Port the Secret Agent is running on.",
                    "type": "string"
                },
                "timeout": {
                    "description": "Timeout in milliseconds.",
                    "type": "integer"
                },
                "tls-ca": {
                    "description": "The path to a trusted CA certificate file in PEM format.",
                    "type": "string"
                },
                "tls-enabled": {
                    "description": "Indicates whether TLS is enabled.",
                    "type": "boolean"
                }
            }
        },
        "model.SeedNode": {
            "description": "SeedNode represents details of a node in the Aerospike cluster.",
            "type": "object",
            "properties": {
                "host-name": {
                    "description": "The host name of the node.",
                    "type": "string"
                },
                "port": {
                    "description": "The port of the node.",
                    "type": "integer"
                },
                "tls-name": {
                    "description": "TLS certificate name used for secure connections (if enabled).",
                    "type": "string"
                }
            }
        },
        "model.Storage": {
            "description": "Storage represents the configuration for a backup storage details.",
            "type": "object",
            "properties": {
                "path": {
                    "description": "The root path for the backup repository.",
                    "type": "string"
                },
                "s3-endpoint-override": {
                    "description": "An alternative endpoint for the S3 SDK to communicate (AWS S3 optional).",
                    "type": "string"
                },
                "s3-log-level": {
                    "description": "The log level of the AWS S3 SDK (AWS S3 optional).",
                    "type": "string",
                    "default": "Fatal"
                },
                "s3-profile": {
                    "description": "The S3 profile name (AWS S3 optional).",
                    "type": "string"
                },
                "s3-region": {
                    "description": "The S3 region string (AWS S3 optional).",
                    "type": "string"
                },
                "type": {
                    "description": "The type of the storage provider (0 - Local, 1 - AWS S3).",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.StorageType"
                        }
                    ]
                }
            }
        },
        "model.StorageType": {
            "description": "StorageType represents the type of the backup storage.",
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "Local",
                "S3"
            ]
        },
...
    },
...
}